pyproject.toml
[project]
name = "Back-End"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "mateusLimadeMacedo", email = "mateuslimamacedo2008@gmail.com"},
    {name = "renan1811", email = "erickechaves@gmail.com"},
    {name = "", email = ""},
    {name = "erickChavesFiamoncini", email = "erickechaves@gmail.com"},
]
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "python-dotenv>=1.1.1",
    "netifaces>=0.11.0",
    "django-filter>=25.1",
    "Pillow>=11.3.0",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python manage.py runserver"
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
user = "python manage.py createsuperuser"
